+++ src/main/java/com/vyshyvan/repositories/militarybasedislocation/MilitaryBaseDislocationRepository.java
@@ -11,7 +11,7 @@ public interface MilitaryBaseDislocationRepository extends JpaRepository<Militar
     List<MilitaryBaseDislocation> findAllByMilitaryBaseId(@Param("id") int id);
 
     @Query("SELECT a FROM MilitaryBaseDislocation a WHERE a.militaryBase.army.id = :id")
-    List<MilitaryBaseDislocation> findAllByMilitaryBaseArmy_Id(@Param("id") int id);
+    List<MilitaryBaseDislocation> findAllByMilitaryBaseArmy_Id(@Param("id") long id);
 
     @Query("SELECT a FROM MilitaryBaseDislocation a WHERE a.militaryBase.division.id = :id")
     List<MilitaryBaseDislocation> findAllByMilitaryBaseDivision_Id(@Param("id") int id);
